% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised.R
\docType{data}
\name{bank}
\alias{bank}
\title{bank dataset}
\format{a data frame with 4520 observations on the following 17 variables.}
\source{
S. Moro, P. Cortez and P. Rita. A Data-Driven Approach to Predict the Success of Bank Telemarketing. Decision Support Systems, Elsevier, 62:22-31, June 2014
}
\usage{
data(bank)
}
\value{
\code{bank}	data frame of data with predictor columns \code{age, job, marital, education, default, balance, housing, loan, contact, day, month, duration, campaign, pdays, previous} and \code{poutcome} with response column \code{y} indicating whether the client has subscribed a term deposit.
}
\description{
the data is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be yes (1) or not (0) subscribed.
}
\details{
The data are from Moro (2014) and taken from the UCI ML website \url{https://archive.ics.uci.edu/ml/datasets/Bank+Marketing}.
}
\keyword{datasets}

