% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{concrete dataset}
\format{a \code{RangedData} instance, 1 row per CpG island.}
\source{
Yeh, I. C. (1998). Modeling of strength of high-performance concrete using artificial neural networks. Cement and Concrete Research, 28(12), 1797-1808. Elsevier.
}
\usage{
data(concrete)
}
\value{
\code{concrete}	data frame of data with predictor columns \code{Cement, BlastFurnaceSlag, FlyAsh, Water, Superplasticizer, CoarseAggregate, FineAggregate} and \code{Age} with response column \code{CompressiveStrength}.
}
\description{
Yeh (1998) describes a collection of data sets from different sources that can be used for modeling the compressive strength of concrete formulations as a functions of their ingredients and age.
}
\details{
The data are from Yeh (1998) and taken from the UCI ML website \url{http://archive.ics.uci.edu/ml/datasets/Concrete+Compressive+Strength}. There are 1030 data points from the UCI website, but the paper states that approximately 1,000 samples were made, but only 727 were analyzed in the source material. It is unclear which samples were excluded.
}
\keyword{datasets}

