% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised.R
\name{plot.glmsel}
\alias{plot.glmsel}
\title{Plot \code{glmsel} objects}
\usage{
\method{plot}{glmsel}(x, ...)
}
\arguments{
\item{x}{an object of class 'glmsel'}

\item{...}{arguments to be passed to and from other methods}
}
\description{
This function produces diagnostic plots of objects of class \code{glmsel} produced by \code{glmsel()} function.
}
\details{
\code{plot} will produce residuals versus fitted values and quantile-quantile plots to diagnose the fit of a generalised linear model. If \code{varsel="lasso"} or \code{varsel="enet"} was selected as arguments in \code{lmsel}, \code{plot(object)} will produce and additional plot with lasso or elastic net variable selection paths.
}
\author{
Michal Knut \email{1105406k@student.gla.ac.uk}.
}

