% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{plot.lmsel}
\alias{plot.lmsel}
\title{Plot \code{lmsel} objects}
\usage{
\method{plot}{lmsel}(x, ...)
}
\arguments{
\item{x}{an object of class 'lmsel'}

\item{...}{arguments to be passed to and from other methods}
}
\description{
This function produces diagnostic plots of objects of class \code{lmsel} produced by \code{lmsel()} function.
}
\details{
\code{plot} will produce residuals versus fitted values and quantile-quantile plots to diagnose the fit of a linear model. If \code{varsel="lasso"} or \code{varsel="enet"} was selected as arguments in \code{lmsel}, \code{plot(object)} will produce and additional plot with lasso or elastic net variable selection paths.
}
\author{
Michal Knut \email{1105406k@student.gla.ac.uk}.
}

