% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{prostate dataset}
\format{a data frame with 97 observations on the following 10 variables.}
\source{
Stamey, T., Kabalin, J., McNeal, J., Johnstone, I., Freiha, F., Redwine, E. and Yang, N (1989) Prostate specific antigen in the diagnosis and treatment of adenocarcinoma of the prostate II. Radical prostatectomy treted patients, Journal of Urology 16: 1076-1083.
}
\usage{
data(prostate)
}
\value{
\code{concrete}	data frame of data with predictor columns \code{lcavol, lweight, age, lbph, svi, lcp, gleason} and \code{pgg45} with response column \code{lpsa} and \code{train} column indicating "training set".
}
\description{
data to examine the correlation between the level of prostate-specific antigen and a number of clinical measures in men who were about to receive a radical prostatectomy.
}
\details{
The last column indicates which 67 observations were used as the "training set" and which 30 as the test set, as described on page 48 in the book.
}
\keyword{datasets}

