% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised.R
\name{summary.glmsel}
\alias{summary.glmsel}
\title{Summary of \code{glmsel} objects}
\usage{
\method{summary}{glmsel}(object, dispersion=NULL, ...)
}
\arguments{
\item{object}{an object of class 'glmsel'}

\item{dispersion}{argument which allows to include dispersion  parameter}

\item{...}{arguments to be passed to and from other methods}
}
\description{
This function produces summary of objects of class \code{glmsel} produced by \code{glmsel()} function.
}
\author{
Michal Knut \email{1105406k@student.gla.ac.uk}.
}

