% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{summary.lmsel}
\alias{summary.lmsel}
\title{Summary of \code{lmsel} objects}
\usage{
\method{summary}{lmsel}(object, ...)
}
\arguments{
\item{object}{an object of class 'lmsel'}

\item{...}{arguments to be passed to and from other methods}
}
\description{
This function produces summary of objects of class \code{lmsel} produced by \code{lmsel()} function.
}
\author{
Michal Knut \email{1105406k@student.gla.ac.uk}.
}

